Cleaning up solution
