+++ Derivco.Orniscient/Derivco.Orniscient.Proxy/Observers/IOrniscientObserver.cs
@@ -1,4 +1,6 @@
-﻿using Orleans;
+﻿using Derivco.Orniscient.Proxy.Grains;
+using Derivco.Orniscient.Proxy.Grains.Models;
+using Orleans;
 
 namespace Derivco.Orniscient.Proxy.Observers
 {
