+++ Derivco.Orniscient/Derivco.Orniscient.TestImplementation/Grains/IFooGrain.cs
@@ -1,19 +1,15 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Derivco.Orniscient.Proxy.Attributes;
 using Orleans;
 
-namespace Derivco.Orniscient.TestImplementation
+namespace Derivco.Orniscient.TestImplementation.Grains
 {
     public interface IFooGrain : IGrainWithGuidKey
     {
         Task KeepAlive();
     }
 
-    [Proxy.Attributes.Orniscient(linkFromType: typeof(SubGrain), linkType: LinkType.SameId,colour:"lightblue" )]
+    [OrniscientGrain(typeof(SubGrain), LinkType.SameId,"lightblue" )]
     public class FooGrain : Grain,IFooGrain {
         public Task KeepAlive()
         {
