+++ Derivco.Orniscient/Derivco.Orniscient.TestImplementation/Grains/ISubGrain.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+using Orleans;
+
+namespace Derivco.Orniscient.TestImplementation.Grains
+{
+    public interface ISubGrain : IGrainWithGuidKey
+    {
+        Task SayHallo();
+    }
+}
