--- Derivco.Orniscient/Derivco.Orniscient.TestImplementation/OrniscientTraceListener.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Derivco.Orniscient.TestImplementation
-{
-    class CustomTraceListener
-    {
-    }
-}
