+++ Derivco.Orniscient/TestHost/Program.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Derivco.Orniscient.TestImplementation;
+using Derivco.Orniscient.TestImplementation.Grains;
 using Orleans;
 
 namespace TestHost
